# Knownmethods optimizations