… test/Microsoft.AspNetCore.Server.Kestrel.Performance/KnownStrings.cs
… test/Microsoft.AspNetCore.Server.KestrelTests/KnownStringsTests.cs
… tools/CodeGenerator/HttpUtilities/HttpUtilities.cs
… tools/CodeGenerator/HttpUtilities/HttpUtilitiesGeneratorHelpers.cs