**Before**
```
 |                  Method |       Mean |    StdDev |
 |------------------------ |----------- |---------- |
 |      GetKnownMethod_GET |  2.1957 ns | 0.0081 ns |
 |  GetKnownMethod_CONNECT | 10.5233 ns | 0.0800 ns |
 |   GetKnownMethod_DELETE |  3.4092 ns | 0.0148 ns |
 |     GetKnownMethod_HEAD |  2.5988 ns | 0.0079 ns |
 |    GetKnownMethod_PATCH |  2.5962 ns | 0.0120 ns |
 |     GetKnownMethod_POST |  2.6509 ns | 0.0857 ns |
 |      GetKnownMethod_PUT |  2.6619 ns | 0.0706 ns |
 |  GetKnownMethod_OPTIONS | 11.4744 ns | 0.0276 ns |
 |    GetKnownMethod_TRACE |  2.5957 ns | 0.0194 ns |
 | GetKnownVersion_HTTP1_1 |  2.7503 ns | 0.0160 ns |
```
**After**
```
 |                  Method |      Mean |    StdDev |
 |------------------------ |---------- |---------- |
 |      GetKnownMethod_GET | 2.1285 ns | 0.0037 ns |
 |  GetKnownMethod_CONNECT | 5.2002 ns | 0.0444 ns |
 |   GetKnownMethod_DELETE | 2.4482 ns | 0.0055 ns |
 |     GetKnownMethod_HEAD | 2.4525 ns | 0.0076 ns |
 |    GetKnownMethod_PATCH | 2.4550 ns | 0.0037 ns |
 |     GetKnownMethod_POST | 2.6373 ns | 0.1053 ns |
 |      GetKnownMethod_PUT | 2.4518 ns | 0.0155 ns |
 |  GetKnownMethod_OPTIONS | 5.2316 ns | 0.0109 ns |
 |    GetKnownMethod_TRACE | 2.4431 ns | 0.0035 ns |
 | GetKnownVersion_HTTP1_1 | 2.6444 ns | 0.0048 ns |
```