@davidfowl 2017-03-02
@arespr can you write a tiny description of the optimization?