@arespr 2017-03-02
I have improved my search alghoritm and
found slightly better function:
```
[MethodImpl(MethodImplOptions.AggressiveInlining)]
private static int GetKnownMethodIndex(ulong value)
{
   var tmp = (int) value & 0x0600000C;

    return ((tmp >> 2) | (tmp >> 23)) & 0x0F;
}
```
Alghoritm is here:
https://github.com/arespr/ConsoleApp1/blob/24ace01d17d1015c43c60d658319f1ed2fc5a043/src/ConsoleApp1/Utils.cs