@arespr 2017-03-02
added unit tests, and repaired performance tests

**Before**
```
 |                  Method |       Mean |    StdErr |    StdDev |     Median |
 |------------------------ |----------- |---------- |---------- |----------- |
 |      GetKnownMethod_GET |  2.0383 ns | 0.0021 ns | 0.0077 ns |  2.0379 ns |
 |  GetKnownMethod_CONNECT | 10.0297 ns | 0.1076 ns | 0.5486 ns |  9.7720 ns |
 |   GetKnownMethod_DELETE |  8.5029 ns | 0.0040 ns | 0.0144 ns |  8.5020 ns |
 |     GetKnownMethod_HEAD |  5.8423 ns | 0.0036 ns | 0.0140 ns |  5.8398 ns |
 |    GetKnownMethod_PATCH |  7.8644 ns | 0.0203 ns | 0.0761 ns |  7.8557 ns |
 |     GetKnownMethod_POST |  4.9119 ns | 0.0023 ns | 0.0087 ns |  4.9110 ns |
 |      GetKnownMethod_PUT |  4.0161 ns | 0.0044 ns | 0.0158 ns |  4.0174 ns |
 |  GetKnownMethod_OPTIONS | 10.6244 ns | 0.0125 ns | 0.0468 ns | 10.6132 ns |
 |    GetKnownMethod_TRACE |  6.8743 ns | 0.0102 ns | 0.0395 ns |  6.8686 ns |
 | GetKnownVersion_HTTP1_1 |  2.6691 ns | 0.0263 ns | 0.1417 ns |  2.5801 ns |
```
**After**
```
 |                  Method |      Mean |    StdErr |    StdDev |
 |------------------------ |---------- |---------- |---------- |
 |      GetKnownMethod_GET | 2.0009 ns | 0.0017 ns | 0.0066 ns |
 |  GetKnownMethod_CONNECT | 4.7313 ns | 0.0015 ns | 0.0057 ns |
 |   GetKnownMethod_DELETE | 5.9907 ns | 0.0596 ns | 0.3154 ns |
 |     GetKnownMethod_HEAD | 5.9507 ns | 0.0014 ns | 0.0053 ns |
 |    GetKnownMethod_PATCH | 4.9168 ns | 0.0489 ns | 0.2017 ns |
 |     GetKnownMethod_POST | 4.7472 ns | 0.0024 ns | 0.0094 ns |
 |      GetKnownMethod_PUT | 6.1486 ns | 0.0731 ns | 0.3014 ns |
 |  GetKnownMethod_OPTIONS | 4.7275 ns | 0.0013 ns | 0.0046 ns |
 |    GetKnownMethod_TRACE | 4.7980 ns | 0.0271 ns | 0.1049 ns |
 | GetKnownVersion_HTTP1_1 | 2.5617 ns | 0.0008 ns | 0.0030 ns |
```