@halter73 2017-03-03
I would like to know we can regenerate this if we change the list of known methods. Can we move the algorithm for https://github.com/arespr/ConsoleApp1/blob/24ace01d17d1015c43c60d658319f1ed2fc5a043/src/ConsoleApp1/Utils.cs to our CodeGenerator project, and have it generate the relevant code in HttpUtilities.cs? You can use the same conventions we use in KnownHeaders and FrameFeatureCollection (e.g. .Generated.cs partial class).

It's important that adding or modifying known methods is drop-in. So that probably involves at least generating the _knownMethods declaration so the array is the correct size, and of course generating GetKnownMethodIndex.