@arespr 2017-03-08
@halter73 I have added HttpUtiliies to CodeGenerator project.