@arespr 2017-03-08
**Dev**

```
 |                  Method |       Mean |    StdDev |
 |------------------------ |----------- |---------- |
 |      GetKnownMethod_GET |  2.1001 ns | 0.0115 ns |
 |  GetKnownMethod_CONNECT | 10.2516 ns | 0.0269 ns |
 |   GetKnownMethod_DELETE |  9.0024 ns | 0.3372 ns |
 |     GetKnownMethod_HEAD |  6.0905 ns | 0.0165 ns |
 |    GetKnownMethod_PATCH |  8.2099 ns | 0.1538 ns |
 |     GetKnownMethod_POST |  5.1370 ns | 0.0133 ns |
 |      GetKnownMethod_PUT |  4.1250 ns | 0.0271 ns |
 |  GetKnownMethod_OPTIONS | 11.3046 ns | 0.5302 ns |
 |    GetKnownMethod_TRACE |  7.1423 ns | 0.0229 ns |
 | GetKnownVersion_HTTP1_1 |  2.9267 ns | 0.0448 ns |
```

**Pr**

```
 |                  Method |      Mean |    StdDev |
 |------------------------ |---------- |---------- |
 |      GetKnownMethod_GET | 2.0916 ns | 0.0065 ns |
 |  GetKnownMethod_CONNECT | 5.0860 ns | 0.0183 ns |
 |   GetKnownMethod_DELETE | 5.0462 ns | 0.0240 ns |
 |     GetKnownMethod_HEAD | 4.9901 ns | 0.0426 ns |
 |    GetKnownMethod_PATCH | 5.0271 ns | 0.0287 ns |
 |     GetKnownMethod_POST | 5.0453 ns | 0.0296 ns |
 |      GetKnownMethod_PUT | 5.0553 ns | 0.0180 ns |
 |  GetKnownMethod_OPTIONS | 5.0340 ns | 0.0458 ns |
 |    GetKnownMethod_TRACE | 5.0139 ns | 0.0228 ns |
 | GetKnownVersion_HTTP1_1 | 2.8913 ns | 0.0071 ns |
```