@arespr 2017-03-09
@Im5tu I can make it public and make shared code or remove id from unit tests.