@halter73 2017-03-12
@arespr The only comments I see from @lm5tu are related to typo the "Numer" which should be changed "Number" instead in a couple places.

Thanks for making these changes. I'll review this more thoroughly Monday, and hopefully we can merge this.